% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_email.R
\name{archive_email}
\alias{archive_email}
\title{Archive all folder contents within email folder}
\usage{
archive_email(delete = TRUE)
}
\arguments{
\item{delete}{Logical indicating if email should be deleted after archiving.
Default: TRUE}
}
\value{
zip file containing any folder contents of the email folder
}
\description{
If an email folder is found within the current working directory,
all the folder contents will be zipped into a single file called email.zip
and the contents deleted.
}
\examples{
\dontrun{
if(interactive()){
 archive_email()
 archive_email(delete=FALSE)
 }
}
}
