% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_mail.R
\name{archive_mail}
\alias{archive_mail}
\title{FUNCTION_TITLE}
\usage{
archive_mail(delete = TRUE)
}
\arguments{
\item{delete}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
