% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topbox.R
\name{botbox}
\alias{botbox}
\title{Function to recode data into Bottom X Box ratings data}
\usage{
botbox(data, bot = 2, minval, replacena = FALSE)
}
\arguments{
\item{data}{A vector object to be converted.  Required.}

\item{bot}{Numeric indicator of the bottom X objects to code as 1. Default: 2.}

\item{minval}{Numeric value of the minimum value (will identify from data if not provided).}

\item{replacena}{Logical value to replace NA with 0. Default: FALSE.}
}
\value{
A converted vector object.
}
\description{
Works on vectors to recode the data into 0,1 values based on Bot1,
Bot2, Bot3, or BotX values.  A minimum value can be indicated (minval) but the
function will find the minimum in the data if not provided.  The replacena
logical allows for replacing any NA with 0, if desired.
}
\examples{
\dontrun{
caddat \%>\%
dplyr::select(q19_1, q19_2) \%>\%
  dplyr::mutate(q19_2 = ifelse(q19_2>=6, NA, q19_2),
                test1 = botbox(q19_1, bot=2),
                test2 = botbox(q19_1),
                test3 = botbox(q19_1, bot=1),
                test4 = botbox(q19_2, bot=2),
                test5 = botbox(q19_2, bot=2, replacena=TRUE)) \%>\%
  summary()
}
}
