% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caddat}
\alias{caddat}
\title{caddat is an example positioning dataset}
\format{
A data frame of 362 observations and 247 variables:
\describe{
\item{respnum}{Respondent ID Number}
\item{speclty}{Specialty Indicator}
\item{q19_1}{Awareness: Product 1}
\item{q19_2}{Awareness: Product 2}
\item{q19_3}{Awareness: Product 3}
\item{q19_4}{Awareness: Product 4}
\item{q19_5}{Awareness: Product 5}
\item{q19_6}{Awareness: Product 6}
\item{q21_01}{Allocation: Product 7}
\item{q21_02}{Allocation: Product 8}
\item{q21_03}{Allocation: Product 9}
\item{q21_04}{Allocation: Product 10}
\item{q21_05}{Allocation: Product 3}
\item{q21_06}{Allocation: Product 4}
\item{q21_07}{Allocation: Product 5}
\item{q21_08}{Allocation: Product 6}
\item{q21_09}{Allocation: Other}
\item{q53_01}{Importance (Unimportant): Attribute 01}
\item{q53_02}{Importance (Unimportant): Attribute 02}
\item{q53_03}{Importance (Unimportant): Attribute 03}
\item{q53_04}{Importance (Unimportant): Attribute 04}
\item{q53_05}{Importance (Unimportant): Attribute 05}
\item{q53_06}{Importance (Unimportant): Attribute 06}
\item{q53_07}{Importance (Unimportant): Attribute 07}
\item{q53_08}{Importance (Unimportant): Attribute 08}
\item{q53_09}{Importance (Unimportant): Attribute 09}
\item{q53_10}{Importance (Unimportant): Attribute 10}
\item{q53_11}{Importance (Unimportant): Attribute 11}
\item{q53_12}{Importance (Unimportant): Attribute 12}
\item{q53_13}{Importance (Unimportant): Attribute 13}
\item{q53_14}{Importance (Unimportant): Attribute 14}
\item{q53_15}{Importance (Unimportant): Attribute 15}
\item{q53_16}{Importance (Unimportant): Attribute 16}
\item{q53_17}{Importance (Unimportant): Attribute 17}
\item{q53_18}{Importance (Unimportant): Attribute 18}
\item{q53_19}{Importance (Unimportant): Attribute 19}
\item{q53_99}{Importance (Unimportant): None}
\item{q54_01}{Importance (Critical): Attribute 01}
\item{q54_02}{Importance (Critical): Attribute 02}
\item{q54_03}{Importance (Critical): Attribute 03}
\item{q54_04}{Importance (Critical): Attribute 04}
\item{q54_05}{Importance (Critical): Attribute 05}
\item{q54_06}{Importance (Critical): Attribute 06}
\item{q54_07}{Importance (Critical): Attribute 07}
\item{q54_08}{Importance (Critical): Attribute 08}
\item{q54_09}{Importance (Critical): Attribute 09}
\item{q54_10}{Importance (Critical): Attribute 10}
\item{q54_11}{Importance (Critical): Attribute 11}
\item{q54_12}{Importance (Critical): Attribute 12}
\item{q54_13}{Importance (Critical): Attribute 13}
\item{q54_14}{Importance (Critical): Attribute 14}
\item{q54_15}{Importance (Critical): Attribute 15}
\item{q54_16}{Importance (Critical): Attribute 16}
\item{q54_17}{Importance (Critical): Attribute 17}
\item{q54_18}{Importance (Critical): Attribute 18}
\item{q54_19}{Importance (Critical): Attribute 19}
\item{q54_99}{Importance (Critical): None}
\item{q55_01_01}{Attribute Association: Product 7 _ Attribute 01}
\item{q55_01_02}{Attribute Association: Product 8 _ Attribute 01}
\item{q55_01_03}{Attribute Association: Product 9 _ Attribute 01}
\item{q55_01_04}{Attribute Association: Product 10 Version A _ Attribute 01}
\item{q55_01_05}{Attribute Association: Product 10 Version B _ Attribute 01}
\item{q55_01_06}{Attribute Association: Product 3 _ Attribute 01}
\item{q55_01_07}{Attribute Association: Product 4 _ Attribute 01}
\item{q55_01_08}{Attribute Association: Product 5 _ Attribute 01}
\item{q55_01_09}{Attribute Association: All are similar _ Attribute 01}
\item{q55_01_10}{Attribute Association: None Perform Well _ Attribute 01}
\item{q55_02_01}{Attribute Association: Product 7 _ Attribute 02}
\item{q55_02_02}{Attribute Association: Product 8 _ Attribute 02}
\item{q55_02_03}{Attribute Association: Product 9 _ Attribute 02}
\item{q55_02_04}{Attribute Association: Product 10 Version A _ Attribute 02}
\item{q55_02_05}{Attribute Association: Product 10 Version B _ Attribute 02}
\item{q55_02_06}{Attribute Association: Product 3 _ Attribute 02}
\item{q55_02_07}{Attribute Association: Product 4 _ Attribute 02}
\item{q55_02_08}{Attribute Association: Product 5 _ Attribute 02}
\item{q55_02_09}{Attribute Association: All are similar _ Attribute 02}
\item{q55_02_10}{Attribute Association: None Perform Well _ Attribute 02}
\item{q55_03_01}{Attribute Association: Product 7 _ Attribute 03}
\item{q55_03_02}{Attribute Association: Product 8 _ Attribute 03}
\item{q55_03_03}{Attribute Association: Product 9 _ Attribute 03}
\item{q55_03_04}{Attribute Association: Product 10 Version A _ Attribute 03}
\item{q55_03_05}{Attribute Association: Product 10 Version B _ Attribute 03}
\item{q55_03_06}{Attribute Association: Product 3 _ Attribute 03}
\item{q55_03_07}{Attribute Association: Product 4 _ Attribute 03}
\item{q55_03_08}{Attribute Association: Product 5 _ Attribute 03}
\item{q55_03_09}{Attribute Association: All are similar _ Attribute 03}
\item{q55_03_10}{Attribute Association: None Perform Well _ Attribute 03}
\item{q55_04_01}{Attribute Association: Product 7 _ Attribute 04}
\item{q55_04_02}{Attribute Association: Product 8 _ Attribute 04}
\item{q55_04_03}{Attribute Association: Product 9 _ Attribute 04}
\item{q55_04_04}{Attribute Association: Product 10 Version A _ Attribute 04}
\item{q55_04_05}{Attribute Association: Product 10 Version B _ Attribute 04}
\item{q55_04_06}{Attribute Association: Product 3 _ Attribute 04}
\item{q55_04_07}{Attribute Association: Product 4 _ Attribute 04}
\item{q55_04_08}{Attribute Association: Product 5 _ Attribute 04}
\item{q55_04_09}{Attribute Association: All are similar _ Attribute 04}
\item{q55_04_10}{Attribute Association: None Perform Well _ Attribute 04}
\item{q55_05_01}{Attribute Association: Product 7 _ Attribute 05}
\item{q55_05_02}{Attribute Association: Product 8 _ Attribute 05}
\item{q55_05_03}{Attribute Association: Product 9 _ Attribute 05}
\item{q55_05_04}{Attribute Association: Product 10 Version A _ Attribute 05}
\item{q55_05_05}{Attribute Association: Product 10 Version B _ Attribute 05}
\item{q55_05_06}{Attribute Association: Product 3 _ Attribute 05}
\item{q55_05_07}{Attribute Association: Product 4 _ Attribute 05}
\item{q55_05_08}{Attribute Association: Product 5 _ Attribute 05}
\item{q55_05_09}{Attribute Association: All are similar _ Attribute 05}
\item{q55_05_10}{Attribute Association: None Perform Well _ Attribute 05}
\item{q55_06_01}{Attribute Association: Product 7 _ Attribute 06}
\item{q55_06_02}{Attribute Association: Product 8 _ Attribute 06}
\item{q55_06_03}{Attribute Association: Product 9 _ Attribute 06}
\item{q55_06_04}{Attribute Association: Product 10 Version A _ Attribute 06}
\item{q55_06_05}{Attribute Association: Product 10 Version B _ Attribute 06}
\item{q55_06_06}{Attribute Association: Product 3 _ Attribute 06}
\item{q55_06_07}{Attribute Association: Product 4 _ Attribute 06}
\item{q55_06_08}{Attribute Association: Product 5 _ Attribute 06}
\item{q55_06_09}{Attribute Association: All are similar _ Attribute 06}
\item{q55_06_10}{Attribute Association: None Perform Well _ Attribute 06}
\item{q55_07_01}{Attribute Association: Product 7 _ Attribute 07}
\item{q55_07_02}{Attribute Association: Product 8 _ Attribute 07}
\item{q55_07_03}{Attribute Association: Product 9 _ Attribute 07}
\item{q55_07_04}{Attribute Association: Product 10 Version A _ Attribute 07}
\item{q55_07_05}{Attribute Association: Product 10 Version B _ Attribute 07}
\item{q55_07_06}{Attribute Association: Product 3 _ Attribute 07}
\item{q55_07_07}{Attribute Association: Product 4 _ Attribute 07}
\item{q55_07_08}{Attribute Association: Product 5 _ Attribute 07}
\item{q55_07_09}{Attribute Association: All are similar _ Attribute 07}
\item{q55_07_10}{Attribute Association: None Perform Well _ Attribute 07}
\item{q55_08_01}{Attribute Association: Product 7 _ Attribute 08}
\item{q55_08_02}{Attribute Association: Product 8 _ Attribute 08}
\item{q55_08_03}{Attribute Association: Product 9 _ Attribute 08}
\item{q55_08_04}{Attribute Association: Product 10 Version A _ Attribute 08}
\item{q55_08_05}{Attribute Association: Product 10 Version B _ Attribute 08}
\item{q55_08_06}{Attribute Association: Product 3 _ Attribute 08}
\item{q55_08_07}{Attribute Association: Product 4 _ Attribute 08}
\item{q55_08_08}{Attribute Association: Product 5 _ Attribute 08}
\item{q55_08_09}{Attribute Association: All are similar _ Attribute 08}
\item{q55_08_10}{Attribute Association: None Perform Well _ Attribute 08}
\item{q55_09_01}{Attribute Association: Product 7 _ Attribute 09}
\item{q55_09_02}{Attribute Association: Product 8 _ Attribute 09}
\item{q55_09_03}{Attribute Association: Product 9 _ Attribute 09}
\item{q55_09_04}{Attribute Association: Product 10 Version A _ Attribute 09}
\item{q55_09_05}{Attribute Association: Product 10 Version B _ Attribute 09}
\item{q55_09_06}{Attribute Association: Product 3 _ Attribute 09}
\item{q55_09_07}{Attribute Association: Product 4 _ Attribute 09}
\item{q55_09_08}{Attribute Association: Product 5 _ Attribute 09}
\item{q55_09_09}{Attribute Association: All are similar _ Attribute 09}
\item{q55_09_10}{Attribute Association: None Perform Well _ Attribute 09}
\item{q55_10_01}{Attribute Association: Product 7 _ Attribute 10}
\item{q55_10_02}{Attribute Association: Product 8 _ Attribute 10}
\item{q55_10_03}{Attribute Association: Product 9 _ Attribute 10}
\item{q55_10_04}{Attribute Association: Product 10 Version A _ Attribute 10}
\item{q55_10_05}{Attribute Association: Product 10 Version B _ Attribute 10}
\item{q55_10_06}{Attribute Association: Product 3 _ Attribute 10}
\item{q55_10_07}{Attribute Association: Product 4 _ Attribute 10}
\item{q55_10_08}{Attribute Association: Product 5 _ Attribute 10}
\item{q55_10_09}{Attribute Association: All are similar _ Attribute 10}
\item{q55_10_10}{Attribute Association: None Perform Well _ Attribute 10}
\item{q55_11_01}{Attribute Association: Product 7 _ Attribute 11}
\item{q55_11_02}{Attribute Association: Product 8 _ Attribute 11}
\item{q55_11_03}{Attribute Association: Product 9 _ Attribute 11}
\item{q55_11_04}{Attribute Association: Product 10 Version A _ Attribute 11}
\item{q55_11_05}{Attribute Association: Product 10 Version B _ Attribute 11}
\item{q55_11_06}{Attribute Association: Product 3 _ Attribute 11}
\item{q55_11_07}{Attribute Association: Product 4 _ Attribute 11}
\item{q55_11_08}{Attribute Association: Product 5 _ Attribute 11}
\item{q55_11_09}{Attribute Association: All are similar _ Attribute 11}
\item{q55_11_10}{Attribute Association: None Perform Well _ Attribute 11}
\item{q55_12_01}{Attribute Association: Product 7 _ Attribute 12}
\item{q55_12_02}{Attribute Association: Product 8 _ Attribute 12}
\item{q55_12_03}{Attribute Association: Product 9 _ Attribute 12}
\item{q55_12_04}{Attribute Association: Product 10 Version A _ Attribute 12}
\item{q55_12_05}{Attribute Association: Product 10 Version B _ Attribute 12}
\item{q55_12_06}{Attribute Association: Product 3 _ Attribute 12}
\item{q55_12_07}{Attribute Association: Product 4 _ Attribute 12}
\item{q55_12_08}{Attribute Association: Product 5 _ Attribute 12}
\item{q55_12_09}{Attribute Association: All are similar _ Attribute 12}
\item{q55_12_10}{Attribute Association: None Perform Well _ Attribute 12}
\item{q55_13_01}{Attribute Association: Product 7 _ Attribute 13}
\item{q55_13_02}{Attribute Association: Product 8 _ Attribute 13}
\item{q55_13_03}{Attribute Association: Product 9 _ Attribute 13}
\item{q55_13_04}{Attribute Association: Product 10 Version A _ Attribute 13}
\item{q55_13_05}{Attribute Association: Product 10 Version B _ Attribute 13}
\item{q55_13_06}{Attribute Association: Product 3 _ Attribute 13}
\item{q55_13_07}{Attribute Association: Product 4 _ Attribute 13}
\item{q55_13_08}{Attribute Association: Product 5 _ Attribute 13}
\item{q55_13_09}{Attribute Association: All are similar _ Attribute 13}
\item{q55_13_10}{Attribute Association: None Perform Well _ Attribute 13}
\item{q55_14_01}{Attribute Association: Product 7 _ Attribute 14}
\item{q55_14_02}{Attribute Association: Product 8 _ Attribute 14}
\item{q55_14_03}{Attribute Association: Product 9 _ Attribute 14}
\item{q55_14_04}{Attribute Association: Product 10 Version A _ Attribute 14}
\item{q55_14_05}{Attribute Association: Product 10 Version B _ Attribute 14}
\item{q55_14_06}{Attribute Association: Product 3 _ Attribute 14}
\item{q55_14_07}{Attribute Association: Product 4 _ Attribute 14}
\item{q55_14_08}{Attribute Association: Product 5 _ Attribute 14}
\item{q55_14_09}{Attribute Association: All are similar _ Attribute 14}
\item{q55_14_10}{Attribute Association: None Perform Well _ Attribute 14}
\item{q55_15_01}{Attribute Association: Product 7 _ Attribute 15}
\item{q55_15_02}{Attribute Association: Product 8 _ Attribute 15}
\item{q55_15_03}{Attribute Association: Product 9 _ Attribute 15}
\item{q55_15_04}{Attribute Association: Product 10 Version A _ Attribute 15}
\item{q55_15_05}{Attribute Association: Product 10 Version B _ Attribute 15}
\item{q55_15_06}{Attribute Association: Product 3 _ Attribute 15}
\item{q55_15_07}{Attribute Association: Product 4 _ Attribute 15}
\item{q55_15_08}{Attribute Association: Product 5 _ Attribute 15}
\item{q55_15_09}{Attribute Association: All are similar _ Attribute 15}
\item{q55_15_10}{Attribute Association: None Perform Well _ Attribute 15}
\item{q55_16_01}{Attribute Association: Product 7 _ Attribute 16}
\item{q55_16_02}{Attribute Association: Product 8 _ Attribute 16}
\item{q55_16_03}{Attribute Association: Product 9 _ Attribute 16}
\item{q55_16_04}{Attribute Association: Product 10 Version A _ Attribute 16}
\item{q55_16_05}{Attribute Association: Product 10 Version B _ Attribute 16}
\item{q55_16_06}{Attribute Association: Product 3 _ Attribute 16}
\item{q55_16_07}{Attribute Association: Product 4 _ Attribute 16}
\item{q55_16_08}{Attribute Association: Product 5 _ Attribute 16}
\item{q55_16_09}{Attribute Association: All are similar _ Attribute 16}
\item{q55_16_10}{Attribute Association: None Perform Well _ Attribute 16}
\item{q55_17_01}{Attribute Association: Product 7 _ Attribute 17}
\item{q55_17_02}{Attribute Association: Product 8 _ Attribute 17}
\item{q55_17_03}{Attribute Association: Product 9 _ Attribute 17}
\item{q55_17_04}{Attribute Association: Product 10 Version A _ Attribute 17}
\item{q55_17_05}{Attribute Association: Product 10 Version B _ Attribute 17}
\item{q55_17_06}{Attribute Association: Product 3 _ Attribute 17}
\item{q55_17_07}{Attribute Association: Product 4 _ Attribute 17}
\item{q55_17_08}{Attribute Association: Product 5 _ Attribute 17}
\item{q55_17_09}{Attribute Association: All are similar _ Attribute 17}
\item{q55_17_10}{Attribute Association: None Perform Well _ Attribute 17}
\item{q55_18_01}{Attribute Association: Product 7 _ Attribute 18}
\item{q55_18_02}{Attribute Association: Product 8 _ Attribute 18}
\item{q55_18_03}{Attribute Association: Product 9 _ Attribute 18}
\item{q55_18_04}{Attribute Association: Product 10 Version A _ Attribute 18}
\item{q55_18_05}{Attribute Association: Product 10 Version B _ Attribute 18}
\item{q55_18_06}{Attribute Association: Product 3 _ Attribute 18}
\item{q55_18_07}{Attribute Association: Product 4 _ Attribute 18}
\item{q55_18_08}{Attribute Association: Product 5 _ Attribute 18}
\item{q55_18_09}{Attribute Association: All are similar _ Attribute 18}
\item{q55_18_10}{Attribute Association: None Perform Well _ Attribute 18}
\item{q55_19_01}{Attribute Association: Product 7 _ Attribute 19}
\item{q55_19_02}{Attribute Association: Product 8 _ Attribute 19}
\item{q55_19_03}{Attribute Association: Product 9 _ Attribute 19}
\item{q55_19_04}{Attribute Association: Product 10 Version A _ Attribute 19}
\item{q55_19_05}{Attribute Association: Product 10 Version B _ Attribute 19}
\item{q55_19_06}{Attribute Association: Product 3 _ Attribute 19}
\item{q55_19_07}{Attribute Association: Product 4 _ Attribute 19}
\item{q55_19_08}{Attribute Association: Product 5 _ Attribute 19}
\item{q55_19_09}{Attribute Association: All are similar _ Attribute 19}
\item{q55_19_10}{Attribute Association: None Perform Well _ Attribute 19}
}
}
\usage{
caddat
}
\description{
Dataset containing attribute associations, importance scores,
awareness values, and allocations from a typical positioning study.
A specialty flag (a grouping variable) is also included.
}
\details{
q19 Awareness uses a large scale were 1 = Not Aware, 2 & 3 are Aware but Not
Used and 4 through 7 are increasing degrees of trial and usage for the following
brands: Product 1, Product 2, Product 3, Product 4, Product 5, Product 6.

q21 Allocation uses the following brands: Product 7, Product 8, Product 9,
Product 10, Product 3, Product 4, Product 5, Product 6, Other.

q53 & q54 Importance Questions, first Unimportant and then Critical.

q55 Attribute Associations follow this convention:  q55_XX_YY where
XX = Attribute and YY = Brand.  XX follows the same numbering as q54 above.
}
\keyword{datasets}
