% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Wrapper function that uses janitor::clean_names on a data set}
\usage{
clean_names(data, data_only = TRUE)
}
\arguments{
\item{data}{The dataframe whose variable names need cleaning.}

\item{data_only}{A logical indicating if only the data should be returned
without the list object. Default: TRUE.}
}
\value{
List containing two tibbles, one for the variable names and one with
the cleaned data, unless data_only=TRUE option which returns just the data tibble.
}
\description{
Convenience wrapper that uses janitor::clean_names to clean up
variable names and returns a list where vars = tibble of original variable
names and data = the original data with cleaned variables names.
}
\examples{
\dontrun{
clean_names(caddat)
clean_names(caddat, data_only=FALSE)
}

}
