% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dictionary.R
\name{create_dictionary}
\alias{create_dictionary}
\title{Wrapper function that uses sjPlot::view_df to create an html data dictionary}
\usage{
create_dictionary(
  data,
  output = NULL,
  type = "default",
  select_vars = everything(),
  ...
)
}
\arguments{
\item{data}{Name of the dataframe from which to build the dictionary.  Required.}

\item{output}{Character string containing the path and file name of the html
file.  If nothing is specified it will default to \"Dictionary.html\" in the
working folder.  Required.}

\item{type}{A setting to indicate if the default output or the simple output
should be returned. Default: 'default'}

\item{select_vars}{List of variables to be included in a dplyr::select statement
just before the dictionary is created to limit variables included. Default: everything()}

\item{...}{Any additional parameters to be passed onto sjPlot::view_df.}
}
\value{
Returns an html file to the location and with the name specified in output
}
\description{
Convenience wrapper that uses sjPlot::view_df to create an html
data dictionary.  sjPlot::view_df is particularly good at pulling out labelled
data that is usually found in Qualtrics datasets downloaded via the API or
from an SPSS file.  The default output will list the variable name, type, label
(if found), number of missing, values, value labels (if found), and frequencies.
Options set to also display limited number of character variables.  A more
basic version is also available that just shows variable labels and values
labels wit the type option.  Output is in html format and is best viewed by
launching outside R/Rstudio.
}
\examples{
\dontrun{
if(interactive()){
 create_directory(test1, output="./output/dictionary.html")
 }
}
}
