% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_frequencies.R
\name{create_frequencies}
\alias{create_frequencies}
\title{Creates a set of frequency tables for each variable in a dataframe}
\usage{
create_frequencies(data, output = NULL, level_cutoff = 55)
}
\arguments{
\item{data}{Name of the dataframe upon which to build the codebook.  Required.}

\item{output}{Character string containing the complete path and file name of
an XLSX file for exporting the resulting dataframe to Excel format.  Optional.}

\item{level_cutoff}{A cut-off value for printing frequencies.  A variable with
this number of values or less will have a frequency table generated, if specified,
otherwise a frequency table will not be generated. Default: 55}
}
\value{
Either a dataframe or exports an Excel file in .xlsx format.
}
\description{
Creates a set of frequency tables for each variable in a dataframe
and exports a list of tables that can be exported to Excel.
}
\examples{
\dontrun{
if(interactive()){
 create_frequencies(test1)
 }
}
}
