% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_id.R
\name{create_id}
\alias{create_id}
\title{Generates a unique ID value for each row in a dataset}
\usage{
create_id(data, var1, var2, name, remove = FALSE)
}
\arguments{
\item{data}{A dataframe containing the original data.  Required.}

\item{var1}{The name of a primary sorting variable, perhaps an existing
identifier upon which the new ID will be based.  Uses tidyeval, so no quotation
is needed.  Optional.}

\item{var2}{The name of a secondary sorting variable, perhaps due to gaps or
issues with the primary sorting variable, upon which the new ID will be based.
Uses tidyeval, so no quotation is needed.  Optional.}

\item{name}{The name of the ID variable in the dataset.  Uses tidyeval, so no
quotation is needed.  Default:  id}

\item{remove}{Logical indicating if the sorting variables (var1, var2) should
be removed.  Default: False.}
}
\value{
A dataframe with the new ID number in the first position.
}
\description{
Generates a unique ID value for each row in a dataset that's an
ordered sequence of numbers ranging from 1 to the number of rows in the data.
This is to replace a potentially arbitrary and non-numerical case ID number
found in the original data.  The dataset is first sorted (ascending) on up to
two other variables (var1 and var2, if specified) to ensure replicability of
ordering.  If duplicates are detected, a warning is issued.  There's an
option to keep the original id variable (default) or to remove the original
id variable (remove = TRUE).  The ID value will be in the first position.
}
\examples{
\dontrun{
if(interactive()){
 dat1 <- data.frame(v1 = c(1,2,3,4,NA,NA,NA,NA), v2 = c(NA,NA,NA,NA,1,2,3,4))
 dat2 <- create_id(dat1)
 dat3 <- create_id(dat1, v1, v2)
 dat4 <- create_id(dat1, v1, v2, name="respnum", remove=TRUE)
 dat1 \%>\%
   create_id(v1, v2)
 }
}
}
