% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qconfig.R
\name{create_qconfig}
\alias{create_qconfig}
\title{Deprecated function to create an invisible qconfig file with credentials necessary for qualtRics}
\usage{
create_qconfig()
}
\description{
All qconfig functions now found in the qkey family of functions (create_qkey, get_qkey, clear_qkey, show_token).
}
