% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Wrapper for a 2-way table from janitor::tabyl}
\usage{
crosstab(data, var1, var2, ...)
}
\arguments{
\item{data}{dataframe object containing the variable}

\item{var1}{the row variable for the crosstab}

\item{var2}{the column variable for the crosstab}

\item{...}{additional arguments passed to janitor::tabyl}
}
\value{
A tabyl object
}
\description{
Generates a 2-way frequency table that is outputed as a janitor::tabyl
object that shows counts, row percents, column percents, and total percents.
Also uses quasiquotation from rlang so that variables need not be quoted.
}
\examples{
\dontrun{
crosstab(caddat, q19_1, q19_2)
}
}
