% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_project.R
\name{delete_project}
\alias{delete_project}
\title{Deletes a project folder and all its contents}
\usage{
delete_project(project, client, root = NULL, override_archive_check = FALSE)
}
\arguments{
\item{project}{A character string with a short moniker to indicate either the
project name or number, e.g. 013-898.  Will be the parent folder, inside of
which all project sub-folders will have been created.  Required.}

\item{client}{An optional character string of the client name.  If specified,
will be a parent folder inside of which the project folder will have been
located, i.e. client/project.}

\item{root}{A optional character string that points to a location for
locating the file path.  This would have been specified in the original set_up
to override the default start_path that may have been set in the Renviron
QPACK_SETUP_ROOT variable.}

\item{override_archive_check}{A logical indicating if the archive check should
be overridden, likely for testing purposes only.  Default:  FALSE.}
}
\value{
Returns noting but deletes folders and the name file.
}
\description{
Deletes the project folder and the descriptor file in the study
folder.  The function needs to detect the presence of the archive file in order
for the delete to be executed.
}
\examples{
\dontrun{
delete_project("013-898")
delete_project(client="Toyota", project="LS2019")
}
}
