% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qconfig.R
\name{delete_qconfig}
\alias{delete_qconfig}
\title{Deprecated function to delete the current .qconfig file}
\usage{
delete_qconfig()
}
\description{
All qconfig functions now found in the qkey family of functions (create_qkey, get_qkey, clear_qkey, show_token).
}
