% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Provides descriptive statistics for a dataframe, similar to Python's Describe}
\usage{
describe(data, round = TRUE)
}
\arguments{
\item{data}{Dataframe to be described.  Required.}

\item{round}{A logical value indicating if output should be rounded for
legibility. Default: TRUE}
}
\value{
Outputs a tibble with the summary statistics
}
\description{
Provides descriptive statistics for a dataframe, similar to Python's
Describe.  Returns the variable name, mean, standard deviation, number of valid
responses, number of missing responses, minimum value, 10\%, 25\%, median,
75\%, 90\%, and maximum value quartiles.  Currently limited to numeric data only;
will exclude any non-numeric data from output and listed as NA in output.
}
\examples{
\dontrun{
if(interactive()){
 describe(caddat)
 }
}
}
