% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_na.R
\name{fix_na}
\alias{fix_na}
\title{Function to fill NA's with either a specified value, the mean, or the
median}
\usage{
fix_na(data, replace = NULL)
}
\arguments{
\item{data}{a vector or dataframe object. Required.}

\item{replace}{either a value or the quoted character string \"mean\". Default: Mean.}
}
\value{
A vector or dataframe object, depending on what was originally specified
}
\description{
Works on vectors or data frames.  If a value is provided, like 0,
then each NA will be replaced with the value.  If replace=\"mean\" is specified,
then all NA's will be replaced by the mean value for that column.  This works
across the dataframe per column.  Median may also be chosen.
}
\examples{
\dontrun{
set.seed(1234)
m <- matrix(sample(c(NA, 1:10), 100, replace = TRUE), 10) \%>\% tibble::as_tibble()
m

m \%>\% fix_na("mean")
m \%>\%
mutate(V1 = fix_na(V1, "mean"))
m \%>\% fix_na(0)
}
}
