% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip transposes a dataframe with t() but outputs a Tibble}
\usage{
flip(data)
}
\arguments{
\item{data}{A dataframe to be transposed.  Required.}
}
\value{
A Tibble of the original dataframe, transposed.
}
\description{
Uses t() to transpose a dataframe but outputs a Tibble instead.
Maintains the row names that a Tibble typically removes.
}
\examples{
\dontrun{
if(interactive()){
 flip(caddat)
 }
}
}
