% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Wrapper for a 1-way table from janitor::tabyl}
\usage{
freq(data, var, ...)
}
\arguments{
\item{data}{Dataframe object containing the variable}

\item{var}{Variable for the frequency}

\item{...}{Additional arguments passed to janitor::tabyl}
}
\value{
A tabyl object
}
\description{
Generates a frequency table that is outputed as a janitor::tabyl
object.  Also uses quasiquotation from rlang so that variables need not be quoted.
}
\examples{
\dontrun{
freq(caddat, speclty)
freq(data=caddat, var=speclty)
freq(caddat$speclty)
}
}
