% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dates.R
\name{get_dates}
\alias{get_dates}
\title{Pulls basic descriptives for the three common date fields from a Qualtrics dataset}
\usage{
get_dates(data)
}
\arguments{
\item{data}{Name of the dataframe that contains the date fields.  Required.}
}
\value{
A dataframe with the name of the date field, the min and max values,
and the percent of days between the two that contain observations
}
\description{
Pulls the min and max values for start_date, end_date, and recorded_date,
if found.  Also calculates the percent of days between the min and max value that
contains a record in the dataframe in order to assess if data collection is ongoing
(higher percentages) or only at specific times (lower percentages).  Will only
pull descriptives for the three typical data fields.
}
\examples{
\dontrun{
if(interactive()){
 get_dates(test3)
 }
}
}
