% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta2}
\alias{meta2}
\title{meta2 is an example dataset used for testing}
\format{
An object of class \code{list} of length 3.
}
\usage{
meta2
}
\description{
Dataset containing a data used for testing.
}
\keyword{datasets}
