% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps_score.R
\name{nps_score}
\alias{nps_score}
\title{Function to calculate NPS score from Likelihood to Recommend data}
\usage{
nps_score(data)
}
\arguments{
\item{data}{The dataframe column containing the LTR data}
}
\value{
Vector of numbers indicating the NPS score
}
\description{
Calculates the NPS Score (100, 0, -100) from vector of data containing
0 to 10 Likelihood to Recommend data.
}
\examples{
\dontrun{
if(interactive()){
 set.seed(423234)
 tibble::tibble(ltr=sample(0:11, 20, replace=TRUE)) \%>\%
   dplyr::mutate(score = nps_score(ltr))
 }
}
}
