% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps_segment.R
\name{nps_segment}
\alias{nps_segment}
\title{nps_segment}
\usage{
nps_segment(data = NULL, variable)
}
\arguments{
\item{data}{Dataframe containing the data.  Required.}

\item{variable}{Variable inside the dataframe that contains the relevant score
to be used to classify into NPS segment.  Must be Likelihood to Recommend type
data where the minimum will be 0 and the maximum will be 10, or NPS type category
data where there are three possible values, -100, 0, or 100.}
}
\value{
Dataframe identical to the input dataframe but with 4 new variables
inserted.
}
\description{
Creates NPS segment variables using either LTR data (ranging from
0 to 10) or from NPS data (-100, 0, +100).  Creates four variables appended to
the dataframe: nps_segment (1= Promoter, 2 = Passive, 3 = Detractor), nps_pro
(1 = Promoter, 0 = Not Promoter), nps_pas (1 = Passive, 0 = Not Passive),
nps_det (1 = Detractor, 0 = Not Detractor).  The new variables will be located
immediately behind the variable that contains the data to be used for classification.
}
\examples{
\dontrun{
if(interactive()){
 dat1 <- qpack::test2 \%>\%
 dplyr::mutate(q4 = dplyr::case_when(q3<= 6 ~ -100,
                                     q3<= 8 ~ 0,
                                     q3>= 9 ~ 100))
 nps_segment(data=dat1, variable=q3)
 nps_segment(data=dat1, variable=q4)
 }
}
}
