% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pii_clean.R
\name{pii_clean}
\alias{pii_clean}
\title{pii_clean}
\usage{
pii_clean(data = NULL, vars = NULL)
}
\arguments{
\item{data}{Name of the dataframe to used.  Required.}

\item{vars}{Vector of column names to be removed, if found.  Optional.}
}
\value{
Dataframe with the variables specified removed but otherwise identical.
}
\description{
Removes common P.I.I. variables from dataset found in Qualtrics
surveys, including IPAddress, LocationLatitude, LocationLongitude, RecipientLastName,
RecipientFirstName, and RecipientEmail.  Additional variables can be dropped
by specifying a vector of column names in the vars option.  Function will only
remove these columns if found; no error or warning is thrown if not found.
}
\examples{
\dontrun{
if(interactive()){
 test1 <- pii_clean(test1)
 test1 <- pii_clean(test1, vars=c("q1", "q2"))
 }
}
}
