% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{qcolors}
\alias{qcolors}
\title{Vector of custom colors}
\format{
An object of class \code{character} of length 4.
}
\usage{
qcolors
}
\description{
Vector of custom colors
}
\examples{
qcolors["my_orange"]

}
\keyword{datasets}
