% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcor_plot.R
\name{qcor_plot}
\alias{qcor_plot}
\title{Original name for qplot_cor function}
\usage{
qcor_plot(
  data = NULL,
  first = NULL,
  names = NULL,
  colors = NULL,
  output = NULL,
  sort = TRUE,
  skeleton = FALSE
)
}
\description{
Original name for the qplot_cor function.  Please use that function
instead.
}
