% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Function to rescale a vector to a new minimum and maximum range}
\usage{
rescale(data, min_val, max_val, flip = FALSE)
}
\arguments{
\item{data}{a vector to be rescaled. Required.}

\item{min_val}{a value to indicate the new minimum value required after rescaling.
Default: 0.}

\item{max_val}{a value to indicate the new maximum value required after rescaling.
Default: 1.}

\item{flip}{a logical indicating if the vector should simply be flipped so that
the max is recoded to the min and the min recoded to the max.  Default: FALSE.}
}
\value{
A vector or dataframe object, depending on what was originally specified
}
\description{
Rescales a vector so that range conforms to a defined minimum
and maximum.  If no minimum or maximum is provided, then the default is to
rescale to a minimum of 0 and a maximum of 1.  This only works on vectors and
not dataframes.
}
\examples{
\dontrun{
set.seed(1234)
m <- data.frame(matrix(sample(c(NA, 1:10), 100, replace = TRUE), 10))

m \%>\%
  dplyr::select(X1) \%>\%
  dplyr::mutate(new = rescale(X1))

m \%>\%
  dplyr::select(X2) \%>\%
  dplyr::mutate(new = rescale(X2, min_val=1, max_val=100))

m \%>\%
  dplyr::select(X3) \%>\%
  dplyr::mutate(new = rescale(X3, min_val=1, max_val=7))

m \%>\%
  dplyr::select(X1) \%>\%
  dplyr::mutate(new = rescale(X1, flip=TRUE))
}
}
