% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_calendar.R
\name{response_calendar}
\alias{response_calendar}
\title{Creates choropleth calendar that shows the relative occurrence of observations
from typical Qualtrics date fields}
\usage{
response_calendar(data, date_var, year = NULL, test = FALSE)
}
\arguments{
\item{data}{data Name of the dataframe that contains the date fields.  Required.}

\item{date_var}{Unquoted name of one of the typical Qualtrics date fields, either
start_date, end_date, or recorded date.  Default: recorded_date.}

\item{year}{Numeric value indicating the year of the calendar to return.  If
none is specified, one choropleth calendar for each year found in the data will
be returned.  Optional.}
}
\value{
calendR graph object
}
\description{
Creates a choropleth calendar using the calendR function.  It shows
a yearly calendar where every day that registers an observation is colored a
shade of red such that the darker the red, the more observations recorded for
that date.
}
\examples{
\dontrun{
if(interactive()){
 response_calendar(test3)
 }
}
}
