% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_colnames.R
\name{set_colnames}
\alias{set_colnames}
\title{Wrapper for a naming columns with pipes from magrittr::set_colnames}
\usage{
set_colnames(...)
}
\arguments{
\item{...}{Arguments passed to magrittr::set_colnames}
}
\value{
Nothing returned, just assigns column names
}
\description{
Convenience wrapper for naming columns with pipes from
magrittr::set_colnames so do not have to write out the function name.
}
\examples{
\dontrun{
caddat \%>\%
  select(1:2) \%>\%
  set_colnames(c("r", "s"))
}
}
