% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{set_up}
\alias{set_up}
\title{Set-up code to create project directories and load needed packages}
\usage{
set_up(
  client = NULL,
  project = NULL,
  task = NULL,
  root = NULL,
  descriptor = NULL,
  folders = NULL,
  pack_load = NULL,
  pack_check = NULL,
  source = NULL,
  qpack = TRUE,
  qmod = FALSE,
  qplot = FALSE
)
}
\arguments{
\item{client}{An optional character string of the client name.  If specified,
will be a parent folder inside of which the project folder will be located,
i.e. client/project.  Optional.}

\item{project}{A character string with a short moniker to indicate either the
project name or number, e.g. 013-898.  Will be the parent folder, inside of
which all project sub-folders will be created.  Recommend keeping short as
possible.  Required.}

\item{task}{A character string with a short moniker to indicate the task or
sub-folder underneath project for situations where there is a need for several
different task-related sub-projects in one project folder, e.g. EDA.  Recommend
keeping as short as possible.  Optional.}

\item{root}{A optional character string that points to a root or starting location
for creating the file path.  This overrides the default root that may be set
in the Renviron QPACK_SETUP_ROOT variable.  Optional.}

\item{descriptor}{A character sting assigning a descriptive name to the project
that is used to create two descriptor files in the folder structure which help
identify what is inside the project folder.  One is placed outside the folder
structure to identify the main folder from others, and the other is placed inside
the folder to act as reminder of the project when inside a file manager.  If
the Renviron variable QPACK_SETUP_EXTERNAL_DESCRIPTOR is set to FALSE, then the
external descriptor file will be suppressed.  Required.}

\item{folders}{A character string or vector of strings containing the
names of folders to be created under the project folder.  This overrides any
folders that may be set in the Renviron QPACK_SETUP_FOLDERS variable.  If not specified
and no QPACK_SETUP_FOLDERS variable is specified in Renviron, then set_up defaults
to creating two folders, data and output.}

\item{pack_load}{A character string or vector of strings containing the names
of the packages to be loaded in the order to be loaded.  Optional.}

\item{pack_check}{A character string or vector of strings containing the names
of the packages to be verified as installed but NOT loaded.  Optional.}

\item{source}{An optional string or vector of strings containing the name of
any R files that should be sourced, perhaps containing project-wide functions
or configuration information.  Each file name should end in .R and a relative
file path from the project directory should be included in each character string,
if not housed inside the project folder, e.g. "function.R" or "./config/functions.R"
if  is housed in a sub-folder named "config".  Optional.}

\item{qpack}{A logical indicating the preference to automatically load the
qpack package after all other packages specified in packages parameter.
Default: TRUE.}

\item{qmod}{A logical indicating the preference to automatically load the
qmod package after all other packages specified in packages parameter.
Default: FALSE.}

\item{qlot}{A logical indicating the preference to automatically load the
qlot package after all other packages specified in packages parameter.
Default: FALSE.}
}
\value{
Creates a folder structure and generates two text files but otherwise
does not return any object.
}
\description{
Part of the qpack package of functions, this sets up project
directories, sets the working directory, loads specified packages, and sources
functions and/or config files, if specified.  Can automatically load a discovered
qkey file.  Option to load qpack package.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 set_up(client="Toyota",
        project="LS2019",
        descriptor="Lead Scoring 2019",
        packages=c("caret", "Cairo", "doParallel", "pROC"),
        source=c("functions.R", "config.R"),
        qpack=TRUE)
 }
}
}
