% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qkey.R
\name{show_token}
\alias{show_token}
\title{FUNCTION_TITLE}
\usage{
show_token(name = NULL)
}
\arguments{
\item{name}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
