% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test2}
\alias{test2}
\title{test2 is an example dataset used for testing}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 200 rows and 4 columns.
}
\usage{
test2
}
\description{
Dataset containing a data used for testing.
}
\keyword{datasets}
