% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test3}
\alias{test3}
\title{test3 is an example dataset used for testing}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4545 rows and 3 columns.
}
\usage{
test3
}
\description{
Dataset containing a data used for testing.
}
\keyword{datasets}
