% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeway.R
\name{threeway}
\alias{threeway}
\title{Performs 3-way crosstab using qpack::crosstab split on a third variable}
\usage{
threeway(data, var1, var2, var3, ...)
}
\arguments{
\item{data}{the dataframe object containing the variable. Required.}

\item{var1}{the row variable for the crosstab. Required.}

\item{var2}{the column variable for the crosstab. Required.}

\item{var3}{the column variable to break the crosstab. Required.}

\item{...}{additional arguments for janitor::tabyl. Optional.}
}
\value{
A tibble object.
}
\description{
Generates a 3-way crosstab, or a qpack::crosstab for each value
of the third variable specified.  Built on janitor::tabyl but outputs a dataframe
to maintain visual elements suitable for exporting (but likely not for the console).
Object shows counts, row percents, column percents, and total percents for each
level of third variable.  Also uses quasiquotation from rlang so that variables
need not be quoted.  var1 and var2 specify the crosstab and var3 specifies the split.
}
\examples{
\dontrun{
output <- threeway(caddat, q19_1, q19_2, speclty)
}
}
