% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topbox.R
\name{topbox}
\alias{topbox}
\title{Function to recode data into Top X Box data}
\usage{
topbox(data, top = 2, maxval, replacena = FALSE)
}
\arguments{
\item{data}{A vector object to be converted.  Required.}

\item{top}{Numeric indicator of the top X objects to code as 1. Default: 2.}

\item{maxval}{Numeric value of the maximum value (will identify from data if not provided).}

\item{replacena}{Logical value to replace NA with 0. Default: FALSE.}
}
\value{
A converted vector object.
}
\description{
Works on vectors to recode the data into 0,1 values based on Top1,
Top2, Top3, or TopX values.  A maximum value can be indicated (maxval) but the
function will find the maximum in the data if not provided.  The replacena
logical allows for replacing any NA with 0, if desired.
}
\examples{
\dontrun{
caddat \%>\%
dplyr::select(q19_1, q19_2) \%>\%
  dplyr::mutate(q19_2 = ifelse(q19_2<=3, NA, q19_2),
                test1 = topbox(q19_1, top=2),
                test2 = topbox(q19_1),
                test3 = topbox(q19_1, top=1),
                test4 = topbox(q19_2, top=2),
                test5 = topbox(q19_2, top=2, replacena=TRUE)) \%>\%
  summary()
}
}
