% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qconfig.R
\name{view_qconfig}
\alias{view_qconfig}
\title{FUNCTION_TITLE}
\usage{
view_qconfig()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
