% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_number.R
\name{week_number}
\alias{week_number}
\title{Function to calculate the week number from data data}
\usage{
week_number(date_column, start_day = 1)
}
\arguments{
\item{date_column}{The dataframe column containing the date data}

\item{start_day}{Parameter passed to lubridate::floor_date parameter week_start
where 1 = Monday, 7 = Sunday, etc. Default: 1}
}
\value{
Vector of numbers indicating the week number
}
\description{
Calculates the week number for a vector of date data.
}
\examples{
\dontrun{
if(interactive()){
 set.seed(43792)
 dat1 <- sample(seq(as.Date('1999/01/01'), as.Date('2000/01/01'), by="day"), 12) \%>\%
   as_tibble() \%>\%
   dplyr::arrange(value) \%>\%
   dplyr::mutate(x=1:n()) \%>\%
   qpack::set_colnames(c("date", "number"))

 dat1 \%>\%
   mutate(week_number = week_number(date))
 }
}
}
