% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xlsx.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Wrapper function for openxlsx that exports a data frame to an XLSX file}
\usage{
write_xlsx(
  data,
  file,
  sheet = NULL,
  over = FALSE,
  overfile = FALSE,
  oversheet = TRUE,
  keepna = FALSE,
  rownames = FALSE
)
}
\arguments{
\item{data}{A dataframe containing to be exported. Required.}

\item{file}{The file path and name, ending in .xlsx, to be written. Required.}

\item{sheet}{The name of the sheet inside the Excel file.  If missing then
the name of the dataframe will be used.  Optional.}

\item{over}{Logical that sets the overwriting defaults to what is the likely
most common values (over = TRUE is the same as overfile = TRUE and oversheet
= TRUE). Default: FALSE.}

\item{overfile}{Logical to indicate if the Excel file should be overwritten
or not. Default: FALSE.}

\item{oversheet}{Logical to indicate if the Sheet inside the file should be
overwritten or not. Default: True.}

\item{keepna}{Logical to set if NA values are kept in the written file
or not. Default: FALSE.}

\item{rownames}{Logical to set if row names should be written to the file
or note.  Default: FALSE.}
}
\value{
This outputs an XLSX file and does not return anything
}
\description{
Takes a dataframe or other table and exports it as an XLSX file.
Sheet name can be specified to add to an existing sheet.  Options include
ability to overwrite the sheet(oversheet), overwrite the file (overfile), or
to use a default (over = TRUE) that overwrites both the sheet and the file.
A keepna option controls whether NA values appear in the sheet.
}
\examples{
\dontrun{
write_xlsx(df, file="C:/TEMP/test.xlsx", over = TRUE, keepna=TRUE)
write_xlsx(df, file="C:/TEMP/test.xlsx", overfile = TRUE,  oversheet=TRUE, keepna=TRUE)
}
}
