% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zap_labels.R
\name{zap_labels}
\alias{zap_labels}
\title{zap_labels}
\usage{
zap_labels(data)
}
\arguments{
\item{data}{Dataframe to be zapped}
}
\value{
Dataframe identical to original but without the labels
}
\description{
Removes labels often found when importing SPSS datasets.  Uses
\code{sjlabelled::remove_all_labels} to remove variable labels and value labels,
also uses \code{haven::zap_formats} and  \code{haven::zap_widths} to remove
extraneous formats and widths missed by \code{sjlabelled::remove_all_labels}.
}
\examples{
\dontrun{
if(interactive()){
 head(test1)
 head(zap_labels(test1))
 }
}
}
