% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_project.R
\name{archive_project}
\alias{archive_project}
\title{Archives a project into a zip file}
\usage{
archive_project(project, client, root = NULL, silent = FALSE)
}
\arguments{
\item{project}{A character string with a short moniker to indicate either the
project name or number, e.g. 013-898.  Will be the parent folder, inside of
which all project sub-folders will have been created.  Required.}

\item{client}{An optional character string of the client name.  If specified,
will be a parent folder inside of which the project folder will have been
located, i.e. client/project.}

\item{root}{An optional character string that points to a location for
locating the file path.  This would have been specified in the original set_up
to override the default start_path that may have been set in the Renviron
QPACK_SETUP_ROOT variable.}

\item{silent}{A logical indicating if the zip output indicating compression progress
should be suppressed or not.  Default: FALSE.}
}
\value{
Creates a zip file in the relevant folder structure containing all project
files and non-empty folders (excluding .git and .Rproj.user) but otherwise does
not return any object.
}
\description{
Uses base zip function to archive all the files and non-empty folders
inside a project folder.  Ignores the .git and .Rproj.user folder.  Creates a zip
file names after the study in the general study directory.
}
\examples{
\dontrun{
archive_project("013-898")
archive_project(client="Toyota", project="LS2019")
}
}
