% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Wrapper function that uses janitor::clean_names on a data set}
\usage{
clean_names(data, create_list = FALSE)
}
\arguments{
\item{data}{The dataframe whose variable names need cleaning.}

\item{create_list}{A logical indicating if list object containing a table of
the transformations and an object containing the transformed data should be
returned (TRUE) or just the data (FALSE) Default: FALSE.}
}
\value{
Either the cleaned dataframe or a list containing two tibbles, one for
the variable names and one with the cleaned data, unless data_only=TRUE option
which returns just the data tibble.
}
\description{
Convenience wrapper that uses janitor::clean_names to clean up
variable names and returns the dataset.  Optional behavior to return a list
where vars = tibble of original variable names and data = the original data
with cleaned variables names.
}
\examples{
\dontrun{
clean_names(caddat)
clean_names(caddat, create_list=TRUE)
}

}
