% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_codebook.R
\name{create_codebook}
\alias{create_codebook}
\title{Creates a codebook or data dictionary for a dataframe}
\usage{
create_codebook(data, output = NULL)
}
\arguments{
\item{data}{Name of the dataframe upon which to build the codebook.  Required.}

\item{output}{Character string containing the complete path and file name of
an XLSX file for exporting the resulting dataframe to Excel format.  Optional.}
}
\value{
Either a dataframe or exports an Excel file in .xlsx format.
}
\description{
Creates a codebook or data dictionary for a dataframe in dataframe
format that can be exported to Excel.  Provides a list of the variables (Column),
randomly pulls an example value for each one (Example), indicates the type of
variable (Type), identifies the number of unique levels (Unique) as well as
the percent of records that are missing (Missing).  Leaves spaces for a
Description and a Note in the file.
}
\examples{
\dontrun{
if(interactive()){
 create_codebook(caddat)
 }
}
}
