% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_na.R
\name{create_na}
\alias{create_na}
\title{Function to fill a value with NA's, defaults to -99}
\usage{
create_na(data, value = -99)
}
\arguments{
\item{data}{a vector or dataframe object. Required.}

\item{value}{the value to be replaced. Default: -99.}
}
\value{
A vector or dataframe object, depending on what was originally specified
}
\description{
Works on vectors or data frames.  If a value is provided, then
each will occurrence with be replaced with an NA.  Defaults to look for the
value -99, the common NA value in Qualtrics survey datasets.
}
\examples{
\dontrun{
set.seed(1234)
m <- matrix(sample(c(-99, 1:10), 100, replace = TRUE), 10) \%>\% tibble::as_tibble()
m
m \%>\% create_na()
}
}
