% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qkey.R
\name{create_qkey}
\alias{create_qkey}
\title{FUNCTION_TITLE}
\usage{
create_qkey(name = NULL, url = NULL, token = NULL)
}
\arguments{
\item{name}{PARAM_DESCRIPTION, Default: NULL}

\item{url}{PARAM_DESCRIPTION, Default: NULL}

\item{token}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
