% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qconfig.R
\name{delete_qconfig}
\alias{delete_qconfig}
\title{Delete the current .qconfig file}
\usage{
delete_qconfig()
}
\value{
Nothing is returned but the qconfig file is deleted
}
\description{
Delete the current .qconfig file, if found.
}
\examples{
\dontrun{
if(interactive()){
 delete_qconfig()
 }
}
}
