% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_labels.R
\name{pull_labels}
\alias{pull_labels}
\title{Generate Variable and Value Labels Exported Data Containing Labels}
\usage{
pull_labels(data, meta_file = NULL)
}
\arguments{
\item{data}{Name of the dataframe after importing.  Required.}

\item{meta_file}{Name of the list output from qualtRics::metadata('surveyID').
Optional.}
}
\value{
A list containing two dataframes; var_labels and val_labels.  var_labels
contains two columns, variable and variable_label and each row lists the name
of a variable in the dataframe along with the variable label found.  val_labels
contains multiple columns.  The first is value, which lists all numeric values
possible in the dataframe and then a column for each variable in the dataset
with the value labels that correspond to the values in the value column.
}
\description{
Generates a list of two dataframes containing variable and value
label information that is found from Qualtrics and exported via qualtRics::fetch_survey
or imported from an SPSS .sav file using haven::read_spss. Changes to API
mean that value labels are no longer exported directly.  It is recommended that
you download an SPSS file and read in using haven::read_spss to preserve these
labels, however, a meta_file option is provided if this is not possible.  Please
be aware that the application of value labels from metadata is not perfect and
may result in unusual or incorrect label application, so use at your own risk.
}
\examples{
\dontrun{
if(interactive()){
 pull_labels(dat1, meta_file=mdat1)
 }
}
}
