% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcor_plot.R
\name{qcor_plot}
\alias{qcor_plot}
\title{FUNCTION_TITLE}
\usage{
qcor_plot(data = NULL, names = NULL, colors = NULL, output = NULL)
}
\arguments{
\item{data}{PARAM_DESCRIPTION, Default: NULL}

\item{names}{PARAM_DESCRIPTION, Default: NULL}

\item{colors}{PARAM_DESCRIPTION, Default: NULL}

\item{output}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
