% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.R
\name{refresh}
\alias{refresh}
\title{Refreshes github install of qpack}
\usage{
refresh(pack_list = "qpack")
}
\arguments{
\item{pack_list}{A character vector with the names of packages to update. Default: qpack.}
}
\value{
Nothing, updates package(s).
}
\description{
Convenience function to perform a remotes::install_github for qpack.
}
\examples{
\dontrun{
refresh()
refresh(c("qpack"))
}
}
