% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refresh}
\alias{refresh}
\title{Refreshes github install of qpack}
\usage{
refresh(dev = FALSE)
}
\arguments{
\item{dev}{A logical to indicate if the development version of qpack should be
installed. Default: FALSE.}
}
\value{
Nothing, updates package(s).
}
\description{
Convenience function to perform a remotes::install_github for qpack.
}
\examples{
\dontrun{
qpack:::refresh()
qpack:::refresh(dev=TRUE)
}
}
\keyword{internal}
