% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qkey.R
\name{show_token}
\alias{show_token}
\title{Show the saved token for a given qkey name}
\usage{
show_token(name = NULL)
}
\arguments{
\item{name}{Character string of the name of a qkey created previously with
create_qkey. Required.}
}
\value{
returns the saved token to the console
}
\description{
Useful when troubleshooting a duplicate name within a key store,
the function shows the saved token for a given name.
}
\examples{
\dontrun{
if(interactive()){
 show_token("Study Name")
 }
}
}
