% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qconfig.R
\name{view_qconfig}
\alias{view_qconfig}
\title{Deprecated function to view and edit the contents of the current .qconfig file}
\usage{
view_qconfig()
}
\description{
All qconfig functions now found in the qkey family of functions (create_qkey, get_qkey, clear_qkey, show_token).
}
