% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qconfig.R
\name{view_qconfig}
\alias{view_qconfig}
\title{View and edit the contents of the current .qconfig file}
\usage{
view_qconfig()
}
\value{
Nothing is returned but the qconfig file is opened
}
\description{
View and edit the contents of the current .qconfig file, if found.
}
\examples{
\dontrun{
if(interactive()){
 view_qconfig()
 }
}
}
